Remove reader functionality
Need for this largely based on misunderstanding. My impression was that leading slash in path was lost when coercing path from string to keyword and coercing it back to string. However, this can be avoided by using `str` instead of `name`:

(name (keyword "/path/to")) => "path/to"
(subs (str (keyword "/path/to")) 1)  => "/path/to"